[feat] heuristics & calculation

+++ Vector2D.h
@@ -4,15 +4,15 @@
 
 struct Vector2D
 {
-    float x = 0; ///< The x-coordinate of the vector.
-    float y = 0; ///< The y-coordinate of the vector.
+    int x = 0; ///< The x-coordinate of the vector.
+    int y = 0; ///< The y-coordinate of the vector.
 
     static const Vector2D Zero; ///< A vector with both coordinates set to zero.
     static const Vector2D One;  ///< A vector with both coordinates set to one.
 
     Vector2D() : x(0), y(0) {}
 
-    Vector2D(float pX, float pY) : x(pX), y(pY) {}
+    Vector2D(int pX, int pY) : x(pX), y(pY) {}
 
     inline void operator+=(Vector2D& right)
     {
@@ -20,7 +20,7 @@ struct Vector2D
         y += right.y;
     }
 
-    inline void operator+=(float length)
+    inline void operator+=(int length)
     {
         x += length;
         y += length;
@@ -32,7 +32,7 @@ struct Vector2D
         y -= right.y;
     }
 
-    inline void operator-=(float length)
+    inline void operator-=(int length)
     {
         x -= length;
         y -= length;
@@ -44,7 +44,7 @@ struct Vector2D
         y *= multiplyer.y;
     }
 
-    inline void operator*=(float length)
+    inline void operator*=(int length)
     {
         x *= length;
         y *= length;
@@ -56,7 +56,7 @@ struct Vector2D
         y /= divider.y;
     }
 
-    inline void operator/=(float length)
+    inline void operator/=(int length)
     {
         x /= length;
         y /= length;
@@ -77,7 +77,7 @@ struct Vector2D
         return { left.x - right.x, left.y - right.y };
     }
 
-    friend Vector2D operator-(Vector2D& left, float right)
+    friend Vector2D operator-(Vector2D& left, int right)
     {
         return { left.x - right, left.y - right };
     }
@@ -97,27 +97,27 @@ struct Vector2D
         return { left.x + right.x, left.y + right.y };
     }
 
-    friend Vector2D operator+(Vector2D& left, float right)
+    friend Vector2D operator+(Vector2D& left, int right)
     {
         return { left.x + right, left.y + right };
     }
 
-    friend Vector2D operator*(float scalar, Vector2D& right)
+    friend Vector2D operator*(int scalar, Vector2D& right)
     {
         return { scalar * right.x, scalar * right.y };
     }
 
-    friend Vector2D operator*(Vector2D& left, float scalar)
+    friend Vector2D operator*(Vector2D& left, int scalar)
     {
         return { left.x * scalar, left.y * scalar };
     }
 
-    friend Vector2D operator/(float scalar, Vector2D& right)
+    friend Vector2D operator/(int scalar, Vector2D& right)
     {
         return { scalar / right.x, scalar / right.y };
     }
 
-    friend Vector2D operator/(Vector2D& left, float scalar)
+    friend Vector2D operator/(Vector2D& left, int scalar)
     {
         return { left.x / scalar, left.y / scalar };
     }
@@ -148,7 +148,7 @@ struct Vector2D
         return left.x * right.x + left.y * right.y;
     }
 
-    inline void Clamp(float minValue, float maxValue)
+    inline void Clamp(int minValue, int maxValue)
     {
         if ((*this).x < minValue)  (*this).x = minValue;
         if ((*this).x > maxValue) (*this).x = maxValue;
@@ -162,7 +162,7 @@ struct Vector2D
         return sqrt(((*this).x - right.x) * ((*this).x - right.x) + ((*this).y - right.y) * ((*this).y - right.y));
     }
 
-    friend Vector2D Clamp(Vector2D temp, float minValue, float maxValue)
+    friend Vector2D Clamp(Vector2D temp, int minValue, int maxValue)
     {
         if (temp.x < minValue)  temp.x = minValue;
         if (temp.x > maxValue)  temp.x = maxValue;
